Move resetManagerDashboard, displayManagerDashboard, resetSearch, resetUserInfo, makeChangeable functions to domUpdates
